+++ src/collectDirectives.ts
@@ -9,37 +9,25 @@ export function collectDirectives(allLines: string[], startIndex: number): stmtL
   const eol = rpgiv.getEOL();
   const settings = rpgiv.getRPGIVFreeSettings();
 
-  // NOTE: Updates various compiler directives to modern versions
-
-  // Find start of comment block by scanning backward
-  let start = startIndex;
-  while (start > 0) {
-    const line = allLines[start-1].trimEnd();
-    if (rpgiv.isEmptyStmt(line) || rpgiv.isDirective(line)) {
-      start--;
-    } else {
-      break;
-    }
-  }
-
-  // Find end of comment block by scanning forward
-  let end = start;
-  while (end < allLines.length) {
-    const line = allLines[end+1].trimEnd();
-     if (rpgiv.isEmptyStmt(line) || rpgiv.isDirective(line)) {
-      end++;
-    } else {
-      break;
-    }
-  }
   const fileInfo = rpgiv.getActiveFileInfo();
   const ext = fileInfo?.extension ? fileInfo.extension.trim() : '';
   const config = rpgiv.getRPGIVFreeSettings();
 
-  for (let i = start; i <= end; i++) {
-    lines.push(convertDirective(allLines[i],ext, config));
-    indexes.push(i);
+  let start = startIndex;
+  // Find end of comment block by scanning forward
+  let end = startIndex;
+  for (let i = start; i < allLines.length; i++) {
+    const line = allLines[i].trimEnd();
+     if (rpgiv.isDirective(line)) {
+      lines.push(convertDirective(allLines[i],ext, config));
+       indexes.push(i);
+       continue;
+    } else if (rpgiv.isSpecEmpty(line) || rpgiv.isComment(line)) {
+      continue
+    }
+    break;
   }
+
 
   return {
     lines,
@@ -53,7 +41,7 @@ function convertDirective(line: string, extension: string,  settings: rpgiv.conf
   if (!rpgiv.isDirective(line)) {
     return line;
   }
-  let enhancedDir = '';
+  let enhancedDir = line;
   const isSQLType = extension?.startsWith('.sql');
   // ...existing code...
   const col = rpgiv.getCol(line, 7, 80);
@@ -65,11 +53,13 @@ function convertDirective(line: string, extension: string,  settings: rpgiv.conf
         dirLower === '/end-free')) {
       enhancedDir = ` // F2FF: Removed deprecated ${dir} statement`
     } else if (settings.replaceCOPYinRPG && dirLower.startsWith('/copy')) {
-      if (isSQLType && settings.replaceCOPYinSQLRPG)
+      if ((isSQLType && settings.replaceCOPYinSQLRPG) || !isSQLType)
      // Replace only the directive, keep spacing after it intact
       enhancedDir = line.replace(/\/copy/i, '/include');
     }
   }
+  // Replace everything before the first '/' (including the '/') with exactly 7 spaces and a '/'
+  enhancedDir = enhancedDir.replace(/^.{0,7}\/\s*/i, '       /');
 
   return enhancedDir;
 }